16-08-2022
About webworkers and an example of webworkers. In the example, we are providing the note to tell the user where the web workers will work. After that we are checking whether the browser support the web workers or not by using a if condition. If the browser support then the javascript will start executing in the background.